fetch and reload items from extracted ItemList viewmodel
